Update patch set 1
Patch Set 1:

As this is based off the shamu commit, I authored it to crpalmer.

Also Fast is reported when plugged into a normal charger, I am asking any user to find out what charger_type is reported when plugged into a quick charger and will submit a PS once that is found.

Patch-set: 1