Update patch set 2
Patch Set 2:

Indeed, I have noticed that charger_type is always set to "Fast" no matter if phone is connected to a standard wall charger or to the computer. Next Friday I will be able to test on a QC 2.0 compliant charger and will check what charger_type is reported.

Patch-set: 2