Update patch set 3
Patch Set 3:

If we are (ever) down to that, the RIL blobs should be specified in m8-common. I put them on proprietary-files.txt so that m8 and m8d can have it split in different vendor folders (even though m8d is happy using those from m8, but we never know).

Patch-set: 3