Update patch set 2
Patch Set 2:

Marc: ULPS and what used to be referred to as Doze in Lollipop is useful for enabling the use of Ambient Display without a massive battery hit though isn't it? Or am I thinking of something else? I remember you spending time to make it work on hammerheadcaf.

Patch-set: 2
Label: Code-Review=0