releasetools: Fix symlink creation in variant script
If destination files exist, variant_script.sh will return 1 to
releasetools.py, which causes the installer to exit with an
error before symlinking all the blobs. Changing 'ln' arguments
to '-sbS .orig' will rename existing destination files while
symlinking all files in /system/blobs/$variant tree to the
/system tree. Also changed the 'find' search path from '.' to
'*' to prevent leading './' from showing in the result, because
that result is used to create source and destination paths for
symlinking and having './' in the middle of concatenated paths
may cause issues.

Change-Id: Iffa66c0a09570c7aa2acf0fdca04fbf54aed8e90