releasetools: Fix symlink creation in variant script
If destination files exist, variant_script.sh will return 1 to
releasetools.py, which causes the installer to exit with an
error before symlinking all the blobs. This patch will remove
any existing destination files. The patch also changes the
'find' search path from '.' to '*' to prevent leading './' from
showing in the result, because that result is used to create
source and destination paths for symlinking and having './' in
the middle of concatenated paths may cause issues.

(Edited: the original patch symlinked everything from
/system/blobs/$variant to /system, renaming any existing
destination files.)

Change-Id: Iffa66c0a09570c7aa2acf0fdca04fbf54aed8e90