Update patch set 2
Patch Set 2:

> (3 comments)
 > 
 > Looks fine. Couple nitpicks on the commit message.
 > 
 > Also - I removed Shane, Jurica and Richard, they're all pretty
 > inactive lately and don't really need to be bothered.

Thanks so much for the guidance, I'm old to OpenSource but new to this
tooling and the standards the team uses. Will this work? If not please 
feel free to be blunt. Standards are good and I've got a thick skin ;-)

Patch-set: 2
Reviewer: Chuck Mattern <16548@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0